7주차 - footer 영역 컴포넌트 분리
