+++ src/components/LinkArea.js
@@ -1,0 +1,14 @@
+const LinkArea = () => {
+  return (
+    <div className="link_area">
+      <a href="./PrivacyPolicy" className="link_text">
+        Privacy Policy
+      </a>
+      <a href="./Faq" className="link_text">
+        FAQ
+      </a>
+    </div>
+  );
+};
+
+export default LinkArea;
