+++ src/constants/ICON_PATH.js
@@ -1,0 +1,13 @@
+import facebookIconPath from '../assets/icon-facebook.svg';
+import twitterIconPath from '../assets/icon-twitter.svg';
+import youtubeIconPath from '../assets/icon-youtube.svg';
+import instargramIconPath from '../assets/icon-instargram.svg';
+
+const ICON_PATH = {
+  facebook: facebookIconPath,
+  twitter: twitterIconPath,
+  youtube: youtubeIconPath,
+  instargram: instargramIconPath,
+};
+
+export default ICON_PATH;
